chore: refactor events commands out of service layer
