+++ src/event_service/commands/__init__.py
@@ -1,2 +1,3 @@
 """Package for all commands."""
 from .contestants_commands import ContestantsCommands
+from .events_commands import EventsCommands
